je veux mourir :'(

+++ ft_atoi.c
@@ -3,10 +3,10 @@
 /*                                                              /             */
 /*   ft_atoi.c                                        .::    .:/ .      .::   */
 /*                                                 +:+:+   +:    +:  +:+:+    */
-/*   By: epoggio <marvin@le-101.fr>                 +:+   +:    +:    +:+     */
+/*   By: epoggio <epoggio@student.42.fr>            +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2018/10/11 18:24:04 by epoggio      #+#   ##    ##    #+#       */
-/*   Updated: 2018/10/11 18:39:30 by epoggio     ###    #+. /#+    ###.fr     */
+/*   Updated: 2018/10/28 21:29:56 by epoggio     ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
@@ -15,8 +15,8 @@
 
 int	ft_atoi(const char *str)
 {
-	long nbr;
-	long neg;
+	long	nbr;
+	int		neg;
 
 	nbr = 0;
 	while (ft_isspace(*str))
@@ -24,6 +24,6 @@ int	ft_atoi(const char *str)
 	neg = (*str == '-') ? -1 : 1;
 	str += (*str == '-' || *str == '+') ? 1 : 0;
 	while (ft_isdigit(*str))
-		nbr = nbr * 10 + *str++ - '0';
-	return ((int)nbr * neg);
+		nbr = (nbr * 10) + (*str++ - '0');
+	return (nbr * neg);
 }
